rename: fileBox와 postBox로 분리
