+++ src/components/ui/box/PostBox.tsx
@@ -1,8 +1,8 @@
 import React, { ComponentProps } from 'react';
-import { PiPaperclipFill } from 'react-icons/pi';
+// import { PiPaperclipFill } from 'react-icons/pi';
 
 import { WithReactChildren } from '@/types/default-interfaces';
-import { ContentFileResponse } from '@/types/page';
+// import { ContentFileResponse } from '@/types/page';
 
 interface IPostBox {
    shadow?: boolean;
@@ -25,32 +25,32 @@ export default function PostBox({
    );
 }
 
-interface FileBoxProps {
-   files: ContentFileResponse[];
-   className?: string;
-   children?: React.ReactNode;
-}
+// interface FileBoxProps {
+//    files: ContentFileResponse[];
+//    className?: string;
+//    children?: React.ReactNode;
+// }
 
-export function FileBox({ files, className, children }: FileBoxProps) {
-   return (
-      <div
-         className={`text-sm px-4 py-5 bg-white rounded-lg shadow-[2px_2px_5px_2px_#00000010] leading-5 transition-opacity animate-fadeIn ${
-            className ?? ''
-         }`}
-      >
-         {files.map((file) => (
-            <a
-               className='flex items-center gap-2'
-               href={file.url}
-               target='_blank'
-               rel='noopener noreferrer'
-               key={file.id}
-            >
-               <PiPaperclipFill style={{ fontSize: '20px' }} />
-               <span className='overflow-hidden text-ellipsis truncate'>{file.originalName}</span>
-            </a>
-         ))}
-         {children}
-      </div>
-   );
-}
+// export default function FileBox({ files, className, children }: FileBoxProps) {
+//    return (
+//       <div
+//          className={`text-sm px-4 py-5 bg-white rounded-lg shadow-[2px_2px_5px_2px_#00000010] leading-5 transition-opacity animate-fadeIn ${
+//             className ?? ''
+//          }`}
+//       >
+//          {files.map((file) => (
+//             <a
+//                className='flex items-center gap-2'
+//                href={file.url}
+//                target='_blank'
+//                rel='noopener noreferrer'
+//                key={file.id}
+//             >
+//                <PiPaperclipFill style={{ fontSize: '20px' }} />
+//                <span className='overflow-hidden text-ellipsis truncate'>{file.originalName}</span>
+//             </a>
+//          ))}
+//          {children}
+//       </div>
+//    );
+// }
